Create Html updated
In this new version of the about me, I tried to make the code a little neater and easier to read. I also added a new section called About me( so the header). I also added a new sections called my hobbies.

Will keep developing this and add the right sections